+++ tensorflow/lite/micro/kernels/arc_mli/scratch_buffers.cc
@@ -67,19 +67,19 @@ static int8_t *scratch_mem[] = {scratch_mem_x, scratch_mem_y, scratch_mem_z};
 static uint32_t scratch_sizes[] = {SCRATCH_MEM_X_SIZE, SCRATCH_MEM_Y_SIZE,
                                    SCRATCH_MEM_Z_SIZE};
 
-void *get_arc_scratch_buffer(uint32_t size) {
+void *get_arc_scratch_buffer(int size) {
   // Function to asign fast memory from one of 3 scratch buffers.
   // Best Fit strategy - memory is allocated from that memory bank that leaves
   // the least unused memory.
   void *buf = NULL;
   int best_mem_idx = -1;
-  uint32_t best_mem_delta = INT_MAX;
+  int best_mem_delta = INT_MAX;
   const int num_mem = sizeof(scratch_mem) / sizeof(scratch_mem[0]);
   // find a local memory that fits the data size.
   for (int mem_idx = 0; mem_idx < num_mem; ++mem_idx) {
     // Best Fit
-    if ((size <= scratch_sizes[mem_idx]) &&
-        (scratch_sizes[mem_idx] - size < best_mem_delta)) {
+    if ((size <= static_cast<int>(scratch_sizes[mem_idx])) &&
+        (static_cast<int>(scratch_sizes[mem_idx]) - size < best_mem_delta)) {
       best_mem_idx = mem_idx;
       best_mem_delta = scratch_sizes[mem_idx] - size;
     }
@@ -92,28 +92,28 @@ void *get_arc_scratch_buffer(uint32_t size) {
   return buf;
 }
 
-void get_arc_scratch_buffer_max_size(uint32_t* size) {
-  uint32_t maxavailable = 0;
+void get_arc_scratch_buffer_max_size(int* size) {
+  int maxavailable = 0;
   const int num_mem = sizeof(scratch_mem) / sizeof(scratch_mem[0]);
   // find the largest available buffer.
   for (int i = 0; i < num_mem; i++) {
-    if (scratch_sizes[i] > maxavailable) {
+    if (static_cast<int>(scratch_sizes[i]) > maxavailable) {
       maxavailable = scratch_sizes[i];
     }
   }
   *size = maxavailable;
 }
 
-void get_arc_scratch_buffer_two_max_sizes(uint32_t* size1, uint32_t* size2) {
-  uint32_t maxavailable = 0;
-  uint32_t secondavail = 0;
+void get_arc_scratch_buffer_two_max_sizes(int* size1, int* size2) {
+  int maxavailable = 0;
+  int secondavail = 0;
   const int num_mem = sizeof(scratch_mem) / sizeof(scratch_mem[0]);
   // find the two largest available buffers.
   for (int i = 0; i < num_mem; i++) {
-    if (scratch_sizes[i] > maxavailable) {
+    if (static_cast<int>(scratch_sizes[i]) > maxavailable) {
       secondavail = maxavailable;
       maxavailable = scratch_sizes[i];
-    } else if (scratch_sizes[i] > secondavail) {
+    } else if (static_cast<int>(scratch_sizes[i]) > secondavail) {
       secondavail = scratch_sizes[i];
     }
   }
